Sepalate MessageList
