+++ src/containers/ChatContainer.js
@@ -1,7 +1,6 @@
 import { connect } from 'react-redux';
 import { join } from '../actions/RoomActions';
 import { set as setRoute } from '../actions/RouteActions';
-import { fetch } from '../actions/MessageActions';
 import { Chat } from '../components/Chat';
 import { bindActions } from './utility';
 
@@ -16,6 +15,5 @@ export const ChatContainer = connect(
     }), bindActions({
         join,
         setRoute,
-        fetch,
     })
 )(Chat);
