+++ src/containers/utility.js
@@ -1,9 +1,15 @@
-export const bindState = (...bindings) => (state) =>
-    bindings.map((key) => [key, state[key]])
-        .reduce((o, v) => {
-            o[v[0]] = v[1];
-            return o;
-        }, {});
+export const bindState = (...keys) => (state) =>
+    keys.reduce((result, key) => {
+        const m = key.match(/^\.{3}(.*)$/);
+
+        if (m) {
+            Object.assign(result, state[m[1]]);
+        } else {
+            result[key] = state[key];
+        }
+
+        return result;
+    }, {});
 
 export const bindActions = (bindings) => (dispatch) =>
      Object.keys(bindings)
